Update 7 rpi segment display project
A key thing I wanted to change was how the code provided in the article I used to put this project together used a bunch of bit arrays and bitwise operations.

Personally I felt that complexity took away the focus on understanding how to use the 7 segement display and I'd spend more time documenting that section than the circuit itself.

So I refactored the code to my own version, which is now included here (after testing it works!).